requirements.txt
streamlit
pandas
google-generativeai
pdfplumber
pytesseract
pdf2image
Pillow
